Move 'ParallelData' into header so Global's destructor works
